requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
PyJWT==2.10.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
